Prepare logical types in validation
