+++ fastavro/logical_writers.py
@@ -1,0 +1,10 @@
+try:
+    from . import _logical_writers
+except ImportError:
+    from . import _logical_writers_py as _logical_writers
+
+LOGICAL_WRITERS = _logical_writers.LOGICAL_WRITERS
+
+__all__ = [
+    'LOGICAL_WRITERS'
+]
