+++ fastavro/write.py
@@ -3,6 +3,7 @@ try:
 except ImportError:
     from . import _write_py as _write
 from . import json_write
+from . import logical_writers
 
 # Private API
 
@@ -11,7 +12,7 @@ writer = _write.writer
 Writer = _write.Writer
 json_writer = json_write.json_writer
 schemaless_writer = _write.schemaless_writer
-LOGICAL_WRITERS = _write.LOGICAL_WRITERS
+LOGICAL_WRITERS = logical_writers.LOGICAL_WRITERS
 
 __all__ = [
     'writer', 'Writer', 'schemaless_writer', 'LOGICAL_WRITERS',
