+++ setup.py
@@ -28,6 +28,7 @@ if not hasattr(sys, 'pypy_version_info'):
         Extension('fastavro._six', ["fastavro/_six" + ext]),
         Extension('fastavro._write', ["fastavro/_write" + ext]),
         Extension('fastavro._validation', ["fastavro/_validation" + ext]),
+        Extension('fastavro._logical_writers', ["fastavro/_logical_writers" + ext]),
     ]
