+++ tests/test_timezone.py
@@ -4,8 +4,8 @@ from io import BytesIO
 import fastavro
 from fastavro.const import MCS_PER_SECOND, MLS_PER_SECOND
 from fastavro import _timezone as tz
-from fastavro._write_py import prepare_timestamp_micros
-from fastavro._write_py import prepare_timestamp_millis
+from fastavro._logical_writers_py import prepare_timestamp_micros
+from fastavro._logical_writers_py import prepare_timestamp_millis
 import pytest
 from .conftest import assert_naive_datetime_equal_to_tz_datetime
